move Problem types to component
