--- packages/ui/src/views/pipeline-edit/components/graph-implementation/types/problem-type.ts
@@ -1,11 +1,0 @@
-export type ProblemSeverity = 'error' | 'warning' | 'info'
-
-export interface Problem<T = unknown> {
-  severity: ProblemSeverity
-  message: string
-  position: {
-    row: number
-    column: number
-  }
-  data?: T
-}
